requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.2
django-appconf==1.0.5
django-compressor==4.1
django-environ==0.9.0
django-multiselectfield==0.1.12
django-on-heroku==1.1.2
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
mysqlclient==2.1.1
phonenumbers==8.13.0
Pillow==9.3.0
psycopg2-binary==2.9.5
pytz==2022.5
rcssmin==1.1.0
rjsmin==1.2.0
sqlparse==0.4.3
whitenoise==6.2.0
